pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Olga Noskova",email = "olga.tolba@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "celery (>=5.6.1,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-django = "^4.11.1"
pytest-cov = "^7.0.0"
flake8 = "^7.3.0"
black = "^25.12.0"
isort = "^7.0.0"
coverage = "^7.13.1"

